requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
pandas==2.2.2
openpyxl==3.1.5
tqdm==4.66.4
requests==2.32.3
python-dotenv==1.0.1
